Increase stream selection dialog item limit (fixed commit)
Bump it from <5 to <=10.  Even 720p monitors display 10 item lists
comfortably and only allowing the display of four streams was arbitrary
and limiting.

Note that there never was an "else" here, so if there's more than 10, it
will just fail to show the list.  This should be corrected in the future
with proper handling of longer lists (or just a retooling of the dialog
code itself to allow scrolling/paging).