Continue to revamp codec checks and support
Created global associative arrays "videoCodecs" and "audioCodecs" to
store codec support bools.

Also altered capabilities.brs to use these and avoid calling
roDeviceInfo functions again when possible (and  added some Debug()
calls to show codec support and strings on the debug console).  This now
allows for every supported codec to be tested against the containers
we're interested in rather than having to rely on docs (e.g. checking if
dts in an mkv container can be decoded) or creating new ifs every time
we want to add a codec.

Adding new codecs for checking is as easy as appending them to their
respective declarations in appMain.brs.