Remove more legacy firmware rev checks
Since this code only works on firmware 7.0+, eliminating more
major/minor checks.  The only one left is anamorphic support, but only
because I can't confirm if it was supported in hardware, firmware, or a
combination of both (the existing code frequently refers to hardware
models adding support for a feature, but checks for firmware revs when
implementing those features, so it's confusing).